Fix images in center tags not rendering by replacing them with centering paragraph tags.
This also requires activating the md_in_html extension as the images are defined with markdown syntax, which is embedded in HTML code. It is unclear how this previously worked without the extension being enabled. It is possible that the center tag, which has been deprecated for a while, was not recognized by the renderer and printed verbatim, then the browser correctly rendered the effects of the tag. With some unknown update of any of the parts involved this is now no longer working. The paragraph solution is the prefered way of centering in markdown in any case.